added loading routes folder for #30; moved api into a separate file for #20
