+++ fabric-starter-client.js
@@ -495,7 +495,7 @@ class FabricStarterClient {
 
     chmodCryptoFolder() {
         //chmodPlus.directory(666, cfg.CRYPTO_CONFIG_DIR, {recursive: true});
-        logger.debug(`Permissions for ${cfg.CRYPTO_CONFIG_DIR} folder are set to 666`);
+        // logger.debug(`Permissions for ${cfg.CRYPTO_CONFIG_DIR} folder are set to 666`);
     }
 
     invalidateChannelsCache(channelId) {
