+++ package.json
@@ -22,6 +22,7 @@
     "express-jwt": "^5.3.1",
     "fabric-ca-client": "^1.4.0",
     "fabric-client": "^1.4.0",
+    "glob": "^7.1.3",
     "jsonwebtoken": "^8.3.0",
     "lodash": "^4.17.11",
     "log4js": "^3.0.6",
